+++ sdk/authorization/arm-authorization/src/authorizationManagementClientContext.ts
@@ -13,7 +13,7 @@ import * as msRest from "@azure/ms-rest-js";
 import * as msRestAzure from "@azure/ms-rest-azure-js";
 
 const packageName = "@azure/arm-authorization";
-const packageVersion = "0.1.0";
+const packageVersion = "8.3.2";
 
 export class AuthorizationManagementClientContext extends msRestAzure.AzureServiceClient {
   credentials: msRest.ServiceClientCredentials;