+++ sdk/cosmosdb/cosmos/src/client/Conflict/index.ts
@@ -1,6 +1,0 @@
-export { Conflict } from "./Conflict";
-export { Conflicts } from "./Conflicts";
-export { ConflictDefinition } from "./ConflictDefinition";
-export { ConflictResponse } from "./ConflictResponse";
-export { ConflictResolutionPolicy } from "./ConflictResolutionPolicy";
-export { ConflictResolutionMode } from "./ConflictResolutionMode";