+++ sdk/cosmosdb/cosmos/src/client/Container/UniqueKeyPolicy.ts
@@ -1,9 +1,0 @@
-/** Interface for setting unique keys on container creation */
-export interface UniqueKeyPolicy {
-  uniqueKeys: UniqueKey[];
-}
-
-/** Interface for a single unique key passed as part of UniqueKeyPolicy */
-export interface UniqueKey {
-  paths: string[];
-}