+++ sdk/cosmosdb/cosmos/src/client/Container/index.ts
@@ -1,5 +1,0 @@
-export { Container } from "./Container";
-export { Containers } from "./Containers";
-export { ContainerDefinition } from "./ContainerDefinition";
-export { ContainerResponse } from "./ContainerResponse";
-export { PartitionKeyRange } from "./PartitionKeyRange";