+++ sdk/cosmosdb/cosmos/src/documents/Document.ts
@@ -1,3 +1,0 @@
-export interface Document {
-  [key: string]: any;
-}