+++ sdk/cosmosdb/cosmos/src/routing/index.ts
@@ -1,5 +1,0 @@
-export * from "./QueryRange";
-export * from "./CollectionRoutingMapFactory";
-export * from "./inMemoryCollectionRoutingMap";
-export * from "./partitionKeyRangeCache";
-export * from "./smartRoutingMapProvider";