+++ sdk/keyvault/keyvault-keys/tests/utils/crypto.ts
@@ -1,0 +1,25 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import { isNode } from "./recorder";
+
+export function stringToUint8Array(str: string): Uint8Array {
+  if (isNode) {
+    return new Uint8Array(Buffer.from(str));
+  } else {
+    const bytes = new Uint8Array(str.length);
+    for (let i = 0; i < str.length; i++) {
+      bytes[i] = str.charCodeAt(i);
+    }
+    return bytes;
+  }
+}
+
+export function uint8ArrayToString(ab: Uint8Array): string {
+  if (isNode) {
+    return Buffer.from(ab).toString("utf-8");
+  } else {
+    const decoder = new TextDecoder("utf-8");
+    return decoder.decode(ab);
+  }
+}